backend/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.6.0
dj-rest-auth==1.1.2
Django==3.1.2
django-allauth==0.43.0
django-cors-headers==3.5.0
djangorestframework==3.12.1
idna==2.10
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
